Unit test for checkout service completed

+++ backend/spring-boot-ecommerce/src/main/java/com/first/springbootecommerce/dao/ProductRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.first.springbootecommerce.Entities.Product;
@@ -17,6 +16,6 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
 	
 	Page<Product> findByNameContaining(@RequestParam("name") String name, Pageable pageable );
 	
-	Page<Product>findById(@RequestParam("id") Long id ,Pageable pageable );
+	Page<Product>findById(@RequestParam("id") Integer id , Pageable pageable);
 
 }
